split up Requirements module
