+++ src/swarm-engine/Swarm/Game/Step/Util/Command.hs
@@ -58,7 +58,7 @@ import Swarm.Game.Universe
 import Swarm.Game.World qualified as W
 import Swarm.Language.Capability
 import Swarm.Language.Pipeline.QQ (tmQ)
-import Swarm.Language.Requirement qualified as R
+import Swarm.Language.Requirements.Type qualified as R
 import Swarm.Language.Syntax
 import Swarm.Language.Text.Markdown qualified as Markdown
 import Swarm.Log
